Move summary to view
