+++ corehq/apps/userreports/views.py
@@ -31,6 +31,8 @@ from corehq.apps.hqwebapp.tasks import send_mail_async
 from corehq.apps.hqwebapp.templatetags.hq_shared_tags import toggle_enabled
 from corehq.apps.reports.daterange import get_simple_dateranges
 from corehq.apps.userreports.specs import FactoryContext
+from corehq.apps.userreports.indicators.factory import IndicatorFactory
+from corehq.apps.userreports.filters.factory import FilterFactory
 from corehq.util import reverse
 from corehq.util.quickcache import quickcache
 from couchexport.export import export_from_tables
@@ -1503,11 +1505,50 @@ class DataSourceSummaryView(BaseUserConfigReportsView):
         return {
             'datasource_display_name': self.config.display_name,
             'datasource_description': self.config.description,
-            'filter_summary': self.config.configured_filter_summary,
-            'indicator_summary': self._add_links_to_output(self.config.indicator_summary),
-            'named_expression_summary': self._add_links_to_output(self.config.named_expression_summary),
-            'named_filter_summary': self._add_links_to_output(self.config.named_filter_summary),
+            'filter_summary': self.configured_filter_summary(),
+            'indicator_summary': self._add_links_to_output(self.indicator_summary()),
+            'named_expression_summary': self._add_links_to_output(self.named_expression_summary()),
+            'named_filter_summary': self._add_links_to_output(self.named_filter_summary()),
         }
+
+    def indicator_summary(self):
+        context = self.config.get_factory_context()
+        wrapped_specs = [
+            IndicatorFactory.from_spec(spec, context).wrapped_spec
+            for spec in self.configured_indicators
+        ]
+        return [
+            {
+                "column_id": wrapped.column_id,
+                "comment": wrapped.comment,
+                "readable_output": wrapped.readable_output(context)
+            }
+            for wrapped in wrapped_specs if wrapped
+        ]
+
+    def named_expression_summary(self):
+        return [
+            {
+                "name": name,
+                "comment": self.config.named_expressions[name].get('comment'),
+                "readable_output": str(exp)
+            }
+            for name, exp in self.config.named_expression_objects.items()
+        ]
+
+    def named_filter_summary(self):
+        return [
+            {
+                "name": name,
+                "comment": self.config.named_filters[name].get('comment'),
+                "readable_output": str(filter)
+            }
+            for name, filter in self.config.named_filter_objects.items()
+        ]
+
+    def configured_filter_summary(self):
+        return str(FilterFactory.from_spec(self.config.configured_filter,
+                                           context=self.config.get_factory_context()))
 
     def _add_links_to_output(self, items):
         def make_link(match):
